+++ addons/test_website/tests/__init__.py
@@ -11,6 +11,7 @@ from . import test_media
 from . import test_multi_company
 from . import test_page
 from . import test_performance
+from . import test_qweb
 from . import test_redirect
 from . import test_reset_views
 from . import test_session
