+++ addons/test_website/tests/template_qweb_test.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-    <template id="website.test_template" name="test template 2">
+    <template id="test_website.test_template" name="test template 2">
 &lt;!DOCTYPE html&gt;
 <html>
     <head>
-        <t t-call-assets="website.test_bundle" t-js="False"/>
+        <t t-call-assets="test_website.test_bundle" t-js="False"/>
         <meta/>
-        <t t-call-assets="website.test_bundle" t-css="False"/>
+        <t t-call-assets="test_website.test_bundle" t-css="False"/>
     </head>
     <body>
         <img src="http://test.external.link/img.png"/>
-        <img src="/website/static/img.png"/>
+        <img src="/test_website/static/img.png"/>
         <a href="http://test.external.link/link">x</a>
         <a href="/web/content/local_link">x</a>
         <span t-attf-style="background-image: url('/web/image/2')" t-att-empty="False">xxx</span>
