+++ addons/website/tests/test_qweb.py
@@ -1,22 +1,13 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from lxml import etree
-import re
-
-from odoo import http, tools
+from odoo import http
 from odoo.addons.base.tests.common import TransactionCaseWithUserDemo
 from odoo.addons.website.tools import MockRequest
-from odoo.modules.module import get_module_resource
 from odoo.tests.common import TransactionCase
 
 
 class TestQweb(TransactionCaseWithUserDemo):
-    def _load(self, module, *args):
-        tools.convert_file(self.cr, 'website',
-                           get_module_resource(module, *args),
-                           {}, 'init', False, 'test')
-
     def test_qweb_post_processing_att(self):
         website = self.env.ref('website.default_website')
         t = self.env['ir.ui.view'].create({
@@ -33,66 +24,6 @@ class TestQweb(TransactionCaseWithUserDemo):
             """
         rendered = self.env['ir.qweb']._render(t.id, {'url': 'http://test.external.img/img2.png'}, website_id=website.id)
         self.assertEqual(rendered.strip(), result.strip())
-
-    def test_qweb_cdn(self):
-        self._load('website', 'tests', 'template_qweb_test.xml')
-
-        website = self.env.ref('website.default_website')
-        website.write({
-            "cdn_activated": True,
-            "cdn_url": "http://test.cdn"
-        })
-
-        demo = self.env['res.users'].search([('login', '=', 'demo')])[0]
-        demo.write({"signature": '''<span class="toto">
-                span<span class="fa"></span><img src="/web/image/1"/>
-            </span>'''})
-
-        demo_env = self.env(user=demo)
-
-        html = demo_env['ir.qweb']._render('website.test_template', {"user": demo}, website_id= website.id)
-        asset_data = etree.HTML(html).xpath('//*[@data-asset-bundle]')[0]
-        asset_xmlid = asset_data.attrib.get('data-asset-bundle')
-        asset_version = asset_data.attrib.get('data-asset-version')
-
-        html = html.strip()
-        html = re.sub(r'\?unique=[^"]+', '', html).encode('utf8')
-
-        attachments = demo_env['ir.attachment'].search([('url', '=like', '/web/assets/%-%/website.test_bundle.%')])
-        self.assertEqual(len(attachments), 2)
-
-        format_data = {
-            "js": attachments[0].url,
-            "css": attachments[1].url,
-            "user_id": demo.id,
-            "filename": "Marc%20Demo",
-            "alt": "Marc Demo",
-            "asset_xmlid": asset_xmlid,
-            "asset_version": asset_version,
-        }
-        self.assertHTMLEqual(html, ("""<!DOCTYPE html>
-<html>
-    <head>
-        <link type="text/css" rel="stylesheet" href="http://test.external.link/style1.css"/>
-        <link type="text/css" rel="stylesheet" href="http://test.external.link/style2.css"/>
-        <link type="text/css" rel="stylesheet" href="http://test.cdn%(css)s" data-asset-bundle="%(asset_xmlid)s" data-asset-version="%(asset_version)s"/>
-        <meta/>
-        <script type="text/javascript" src="http://test.external.link/javascript1.js"></script>
-        <script type="text/javascript" src="http://test.external.link/javascript2.js"></script>
-        <script type="text/javascript" src="http://test.cdn%(js)s" data-asset-bundle="%(asset_xmlid)s" data-asset-version="%(asset_version)s"></script>
-    </head>
-    <body>
-        <img src="http://test.external.link/img.png" loading="lazy"/>
-        <img src="http://test.cdn/website/static/img.png" loading="lazy"/>
-        <a href="http://test.external.link/link">x</a>
-        <a href="http://test.cdn/web/content/local_link">x</a>
-        <span style="background-image: url(&#39;http://test.cdn/web/image/2&#39;)">xxx</span>
-        <div widget="html"><span class="toto">
-                span<span class="fa"></span><img src="http://test.cdn/web/image/1" loading="lazy">
-            </span></div>
-        <div widget="image"><img src="http://test.cdn/web/image/res.users/%(user_id)s/avatar_1920/%(filename)s" class="img img-fluid" alt="%(alt)s" loading="lazy"/></div>
-    </body>
-</html>""" % format_data).encode('utf8'))
 
 
 class TestQwebProcessAtt(TransactionCase):
